+++ src/pages/lesson-planning.tsx
@@ -2,7 +2,6 @@ import { FC } from "react";
 import { NextPage, GetStaticProps, GetStaticPropsResult } from "next";
 import { PortableText } from "@portabletext/react";
 
-import useTrackPageView from "../hooks/useTrackPageView";
 import CMSClient from "../node-lib/cms";
 import { PlanningPage, PortableTextJSON } from "../common-lib/cms-types";
 import { decorateWithIsr } from "../node-lib/isr";
@@ -165,8 +164,6 @@ const LessonElementsCard: FC<CardProps> = (props) => (
 );
 
 const PlanALesson: NextPage<PlanALessonProps> = ({ pageData }) => {
-  useTrackPageView({ pageName: "Plan a Lesson" });
-
   return (
     <Layout seoProps={getSeoProps(pageData.seo)} $background={"white"}>
       <MaxWidth $pt={[72, 80, 80]}>